Fix size setup ioctls for small devices
For devices smaller than 4096 the current code results in a blockdevice
size of 0.

For devices with a different blocksize than 4096 the code potentially
leads to more dropped blocks at the end of the file than necessary. For
example a device with size 6144 would have only 4096 bytes available on
client side even if the blocksize is set to 512.

This patch tries to use the user-given blocksize if possible. Otherwise
it falls back to using 4096.

Signed-off-by: Markus Pargmann <mpa@pengutronix.de>